Open the ASTValidator API to external authors
